#161: use deleteReply to consolidate viewcollectartifact flow
