+++ source/buttons/_buttonDictionary.js
@@ -9,6 +9,7 @@ for (const file of [
 	"buylife.js",
 	"buyscouting.js",
 	"challenges.js",
+	"collectartifact.js",
 	"confirmmove.js",
 	"deploy.js",
 	"elementresearch.js",
@@ -33,7 +34,6 @@ for (const file of [
 	"tinker.js",
 	"trainingdummy.js",
 	"upgrade.js",
-	"viewcollectartifact.js",
 	"viewgearcollector.js",
 	"viewgearsales.js",
 	"viewrepairs.js",
